Sample scarcity problem.

+++ src/lidar_sim/include/cuda_funcs.h
@@ -7,7 +7,6 @@
 
 #include <iostream>
 
-
 struct Obsp {
     float x = 0.0;
     float y = 0.0;
@@ -25,7 +24,7 @@ __host__ void copyRawSegs(const float* const host_segs, size_t byte_num);
  * @param segments constant memory片段信息
  * @param flags 共享内存flags, x, y是观测点位置
  */
-__device__ void initialize(const float* const segments, const Eigen::Vector2d* const obs, int id, bool* flags);
+__device__ void initialize(const float* const segments, const Eigen::Vector2f* const obs, int id, bool* flags);
 
 /**
  * @brief 并行的z buffer, 渲染深度图
@@ -33,16 +32,15 @@ __device__ void initialize(const float* const segments, const Eigen::Vector2d* c
  * @param range 是共享内存的range，只有Z buffer全部计算完（并且转换为int） 才可以开始覆盖
  */
 __device__ void singleSegZbuffer(
-    const Eigen::Vector2d& p1, const Eigen::Vector2d& p2, const Obsp* const ptcls, 
-    const int s_id, const int e_id, const int range_num, const double ang_incre, int* range
+    const Eigen::Vector2f& p1, const Eigen::Vector2f& p2, const Obsp* const ptcls, 
+    const int s_id, const int e_id, const int range_num, const float ang_incre, int* range
 );
 
 /// 输入原始的segment
 __global__ void particleFilter(
     const Obsp* const ptcls,
-    // const float* const raw_segs,
     const float* const ref, float* weights,
-    const double ang_min, const double ang_incre, const int range_num, 
+    const float ang_min, const float ang_incre, const int range_num, 
     const int full_rnum, const bool single_flag = false
 );
